+++ inst/NEWS.md
@@ -1,13 +1,15 @@
-Changes in Version 2.2.0 (2017-11-17)
+Changes in Version 2.3.0 (2017-11-20)
 --------------------------------------------------------
 
 NEW FEATURES
 
+* `AsDate` is now stricter when parsing formats without days.
+  Previously, the function could be too aggressive in certain cases and
+  parse non-date input as dates.
 * `AsDate` gains additional support for date ranges/periods with
 full name, abbreviated name, or numeric months; additional sepators
 allowed between parts of the date and separating the dates; and
 support for two or four digit years.
-
 * `AsDate` now always returns an object of class "Date" and 
 `AsDateTime` always returns an object of class "POSIXct".  If dates
 (without times) are supplied to `AsDateTime`, they will be parsed by